Add fade transitions across scenes
1 sec fade transition. Output video does not work in vlc for some
reason. In some videos, the sync issues across the video is a bit
more prominent. But overall looks good.

Kiran's homelabbing talk has a lot of back and forth between the
slides and objects he showed live.  These are shown much better
in the video now. Kiran used his laptop cam to show the live
audience some features, But those are mostly much better visible
in the fullscreen video.

The starting and ending fullscreen video of the user serves to
adequately highlight the speaker as well.